requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.2.0
execnet==2.0.2
fasteners==0.19
filelock==3.13.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
parameterized==0.9.0
parse==1.20.0
parse-type==0.6.2
pdbp==1.5.0
platformdirs==4.1.0
pluggy==1.3.0
py==1.11.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pynose==1.4.8
pyotp==2.9.0
PySocks==1.7.1
pytest==7.4.4
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
sbvirtualdisplay==1.3.0
selenium==4.16.0
seleniumbase==4.22.6
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.42.0
wsproto==1.2.0
yarl==1.9.4
